apps/backend/requirements.txt
fastapi
uvicorn[standard]
pipenv
pytest
mypy
types-psycopg2
typing-extensions
pydantic
pydantic-settings
psycopg2-binary
sqlalchemy
alembic
passlib
bcrypt
